Create p5 sketch G Major Sound Yellow
p5 code that helps the user associate the piano chords for the G major. Whenever a note within the G major chord progression is played, the p5 code should be able to hear the sound, and then the p5 screen turns into a yellow circle animation that loops when a note in G Major is pressed. As this is an HTML script, you will need to post this in a notes application, save it as .htm, and then open it in a browser.